staking passing, low test coverage
